backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.2
whitenoise>=6.7
dj-database-url>=2.1
python-dotenv>=1.0
